+++ musique/interpreter/builtin_functions.cc
@@ -1,7 +1,6 @@
 #include <musique/algo.hh>
 #include <musique/interpreter/env.hh>
 #include <musique/guard.hh>
-#include <musique/interpreter/incoming_midi.hh>
 #include <musique/interpreter/interpreter.hh>
 #include <musique/try.hh>
 
@@ -11,13 +10,6 @@
 #include <unordered_set>
 #include <chrono>
 #include <thread>
-
-void Interpreter::register_callbacks()
-{
-	assert(callbacks == nullptr, "This field should be uninitialized");
-	callbacks = std::make_unique<Interpreter::Incoming_Midi_Callbacks>();
-	callbacks->add_callbacks(*midi_connection, *this);
-}
 
 /// Check if type has index method
 template<typename T>
@@ -67,8 +59,8 @@ static inline std::optional<Error> create_chord(std::vector<Note> &chord, Interp
 		}
 
 		if (auto arg_chord = get_if<Chord>(arg)) {
-			std::ranges::copy_if(
-				arg_chord->notes,
+			std::copy_if(
+				arg_chord->notes.begin(), arg_chord->notes.end(),
 				std::back_inserter(chord),
 				[](Note const& n) { return n.base.has_value(); }
 			);
@@ -80,7 +72,7 @@ static inline std::optional<Error> create_chord(std::vector<Note> &chord, Interp
 			continue;
 		}
 
-		assert(false, "this type is not supported inside chord"); // TODO(assert)
+		ensure(false, "this type is not supported inside chord"); // TODO(assert)
 	}
 
 	return {};
@@ -90,7 +82,7 @@ static inline std::optional<Error> create_chord(std::vector<Note> &chord, Interp
 template<auto Mem_Ptr>
 static Result<Value> ctx_read_write_property(Interpreter &interpreter, std::vector<Value> args)
 {
-	assert(args.size() <= 1, "Ctx get or set is only supported (wrong number of arguments)"); // TODO(assert)
+	ensure(args.size() <= 1, "Ctx get or set is only supported (wrong number of arguments)"); // TODO(assert)
 
 	using Member_Type = std::remove_cvref_t<decltype(std::declval<Context>().*(Mem_Ptr))>;
 
@@ -98,7 +90,7 @@ static Result<Value> ctx_read_write_property(Interpreter &interpreter, std::vect
 		return Number(interpreter.context_stack.back().*(Mem_Ptr));
 	}
 
-	assert(std::holds_alternative<Number>(args.front().data), "Ctx only holds numeric values");
+	ensure(std::holds_alternative<Number>(args.front().data), "Ctx only holds numeric values");
 
 	if constexpr (std::is_same_v<Member_Type, Number>) {
 		interpreter.context_stack.back().*(Mem_Ptr) = std::get<Number>(args.front().data);
@@ -118,7 +110,7 @@ static Result<Array> into_flat_array(Interpreter &interpreter, std::span<Value> 
 	for (auto &arg : args) {
 		std::visit(Overloaded {
 			[&target](Array &&array) -> std::optional<Error> {
-				std::ranges::move(array.elements, std::back_inserter(target.elements));
+				std::move(array.elements.begin(), array.elements.end(), std::back_inserter(target.elements));
 				return {};
 			},
 			[&target, &interpreter](Block &&block) -> std::optional<Error> {
@@ -276,7 +268,7 @@ static std::optional<Error> action_play(Interpreter &i, Value v)
 template<With_Index_Operator Container = std::vector<Value>>
 static inline Result<Value> builtin_play(Interpreter &i, Container args)
 {
-	Try(ensure_midi_connection_available(i, Midi_Connection_Type::Output, "play"));
+	Try(ensure_midi_connection_available(i, "play"));
 	auto previous_action = std::exchange(i.default_action, action_play);
 	i.context_stack.push_back(i.context_stack.back());
 
@@ -298,12 +290,12 @@ static inline Result<Value> builtin_play(Interpreter &i, Container args)
 
 /// Play first argument while playing all others
 static Result<Value> builtin_par(Interpreter &i, std::vector<Value> args) {
-	Try(ensure_midi_connection_available(i, Midi_Connection_Type::Output, "par"));
+	Try(ensure_midi_connection_available(i, "par"));
 
-	assert(args.size() >= 1, "par only makes sense for at least one argument"); // TODO(assert)
+	ensure(args.size() >= 1, "par only makes sense for at least one argument"); // TODO(assert)
 	if (args.size() == 1) {
 		auto chord = get_if<Chord>(args.front());
-		assert(chord, "Par expects music value as first argument"); // TODO(assert)
+		ensure(chord, "Par expects music value as first argument"); // TODO(assert)
 		Try(i.play(std::move(*chord)));
 		return Value{};
 	}
@@ -311,7 +303,7 @@ static Result<Value> builtin_par(Interpreter &i, std::vector<Value> args) {
 	// Create chord that should sustain during playing of all other notes
 	auto &ctx = i.context_stack.back();
 	auto chord = get_if<Chord>(args.front());
-	assert(chord, "par expects music value as first argument"); // TODO(assert)
+	ensure(chord, "par expects music value as first argument"); // TODO(assert)
 
 	std::for_each(chord->notes.begin(), chord->notes.end(), [&](Note &note) { note = ctx.fill(note); });
 
@@ -674,7 +666,7 @@ static Result<Value> builtin_update(Interpreter &i, std::vector<Value> args)
 /// Return typeof variable
 static Result<Value> builtin_typeof(Interpreter&, std::vector<Value> args)
 {
-	assert(args.size() == 1, "typeof expects only one argument"); // TODO(assert)
+	ensure(args.size() == 1, "typeof expects only one argument"); // TODO(assert)
 	return Symbol(type_name(args.front()));
 }
 
@@ -701,7 +693,7 @@ static Result<Value> builtin_shuffle(Interpreter &i, std::vector<Value> args)
 {
 	static std::mt19937 rnd{std::random_device{}()};
 	auto array = Try(flatten(i, std::move(args)));
-	std::ranges::shuffle(array, rnd);
+	std::shuffle(array.begin(), array.end(), rnd);
 	return array;
 }
 
@@ -709,7 +701,7 @@ static Result<Value> builtin_shuffle(Interpreter &i, std::vector<Value> args)
 static Result<Value> builtin_permute(Interpreter &i, std::vector<Value> args)
 {
 	auto array = Try(flatten(i, std::move(args)));
-	std::ranges::next_permutation(array);
+	std::next_permutation(array.begin(), array.end());
 	return array;
 }
 
@@ -717,7 +709,7 @@ static Result<Value> builtin_permute(Interpreter &i, std::vector<Value> args)
 static Result<Value> builtin_sort(Interpreter &i, std::vector<Value> args)
 {
 	auto array = Try(flatten(i, std::move(args)));
-	std::ranges::sort(array);
+	std::sort(array.begin(), array.end());
 	return array;
 }
 
@@ -725,7 +717,7 @@ static Result<Value> builtin_sort(Interpreter &i, std::vector<Value> args)
 static Result<Value> builtin_reverse(Interpreter &i, std::vector<Value> args)
 {
 	auto array = Try(flatten(i, std::move(args)));
-	std::ranges::reverse(array);
+	std::reverse(array.begin(), array.end());
 	return array;
 }
 
@@ -733,7 +725,7 @@ static Result<Value> builtin_reverse(Interpreter &i, std::vector<Value> args)
 static Result<Value> builtin_min(Interpreter &i, std::vector<Value> args)
 {
 	auto array = Try(deep_flat(i, args));
-	if (auto min = std::ranges::min_element(array); min != array.end())
+	if (auto min = std::min_element(array.begin(), array.end()); min != array.end())
 		return *min;
 	return Value{};
 }
@@ -742,7 +734,7 @@ static Result<Value> builtin_min(Interpreter &i, std::vector<Value> args)
 static Result<Value> builtin_max(Interpreter &i, std::vector<Value> args)
 {
 	auto array = Try(deep_flat(i, args));
-	if (auto max = std::ranges::max_element(array); max != array.end())
+	if (auto max = std::max_element(array.begin(), array.end()); max != array.end())
 		return *max;
 	return Value{};
 }
@@ -911,28 +903,6 @@ static Result<Value> builtin_note_off(Interpreter &i, std::vector<Value> args)
 	};
 }
 
-/// Add handler for incoming midi messages
-static Result<Value> builtin_incoming(Interpreter &i, std::vector<Value> args)
-{
-	if (auto a = match<Symbol, Function>(args)) {
-		auto& [symbol, fun] = *a;
-		if (symbol == "note_on" || symbol == "noteon") {
-			i.callbacks->note_on = std::move(args[1]);
-		} else if (symbol == "note_off" || symbol == "noteoff") {
-			i.callbacks->note_off = std::move(args[1]);
-		} else {
-
-		}
-		return Value{};
-	}
-
-	return errors::Unsupported_Types_For {
-		.type = errors::Unsupported_Types_For::Function,
-		.name = "incoming",
-		.possibilities = { "(symbol, function) -> nil" }
-	};
-}
-
 /// Interleaves arguments
 static Result<Value> builtin_mix(Interpreter &i, std::vector<Value> args)
 {
@@ -999,7 +969,6 @@ void Interpreter::register_builtin_functions()
 	global.force_define("for",            builtin_for);
 	global.force_define("hash",           builtin_hash);
 	global.force_define("if",             builtin_if);
-	global.force_define("incoming",       builtin_incoming);
 	global.force_define("instrument",     builtin_program_change);
 	global.force_define("len",            builtin_len);
 	global.force_define("max",            builtin_max);