+++ musique/value/block.cc
@@ -16,7 +16,7 @@ static inline std::optional<Error> guard_index(unsigned index, unsigned size)
 // TODO Add memoization
 Result<Value> Block::index(Interpreter &i, unsigned position) const
 {
-	assert(parameters.size() == 0, "cannot index into block with parameters (for now)");
+	ensure(parameters.size() == 0, "cannot index into block with parameters (for now)");
 	if (body.type != Ast::Type::Sequence) {
 		Try(guard_index(position, 1));
 		return i.eval((Ast)body);